HACK! rs485: add delay after write to simulate real RS485 tty
In Linux, when you write(2) to a real tty, it only unblocks when the
data is actually written to the bus. This behavior is important for the
bacnet-stack code.

However, write(2)'ing to a pts tty doesn't block at all, and it seems
there's no way to alter this behaviour. The only solution is to add an
artificial delay right after write(2).

WARNING: The code uses a constant value of 1040 to represent the time to
transfer one byte ((8+2) bits * 104us) at a 9600 baud rate (1/9600
seconds). It won't work for a baud rate of 38400.

+ add random delay to make everything work