Don't unconditionally send Simple-ACK for WP with the same value
ASHRAE 135.1-2020 7.2.3 test checks that all properties that are marked
in EPICS as read-only are acutally read-only. To do so, each marked
property should be written a) with the same value as it was at the
beginning of the test b) with some other appropriate for this property
value.

In both cases IUT should return some error (e.g. WRITE_ACCESS_DENIED),
however without this change IUT will return Simple-ACK as an answer for
WriteProperty request to object-name of any existing object with the
same value as the property already contains, even if the object itself
doesn't support changing its object-name.